+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add teams
   - A user can join multiple teams
   - Contacts/Contact Groups/Announcements are seperated by team
+- Add permissions and roles
 
 ### Changed
 - Move `User` class into `Models` directory